Explicit cast

+++ test/default/aead_aes256gcm.c
@@ -3139,7 +3139,7 @@ tv(void)
         if (memcmp(ciphertext, expected_ciphertext, ciphertext_len) != 0) {
             printf("Encryption of test vector #%u failed\n", (unsigned int) i);
             hex = (char *) sodium_malloc(found_ciphertext_len * 2 + 1);
-            sodium_bin2hex(hex, found_ciphertext_len * 2 + 1,
+            sodium_bin2hex(hex, (size_t) found_ciphertext_len * 2 + 1,
                            ciphertext, ciphertext_len);
             printf("Computed: [%s]\n", hex);
             sodium_free(hex);
